refactor handling of item score change after submitting answer
