+++ src/app/modules/item/pages/item-content/item-content.component.html
@@ -17,6 +17,7 @@
         [view]="taskView"
         (viewChange)="taskViewChange.emit($event)"
         (tabsChange)="taskTabsChange.emit($event)"
+        (scoreChange)="scoreChange.emit($event)"
       ></alg-item-display>
 
       <alg-sub-skills [itemData]="itemData" *ngIf="itemData.item.type === 'Skill'"></alg-sub-skills>
