+++ src/app/modules/item/services/item-task-answer.service.ts
@@ -15,8 +15,6 @@ import {
 import { SECONDS } from 'src/app/shared/helpers/duration';
 import { errorIsHTTPForbidden } from 'src/app/shared/helpers/errors';
 import { repeatLatestWhen } from 'src/app/shared/helpers/repeatLatestWhen';
-import { isItemInfo, itemInfo } from 'src/app/shared/models/content/item-info';
-import { CurrentContentService } from 'src/app/shared/services/current-content.service';
 import { AnswerTokenService } from '../http-services/answer-token.service';
 import { Answer, CurrentAnswerService } from '../http-services/current-answer.service';
 import { GradeService } from '../http-services/grade.service';
@@ -28,6 +26,9 @@ const answerAndStateSaveInterval = 1*SECONDS;
 export class ItemTaskAnswerService implements OnDestroy {
   private errorSubject = new Subject<Error>();
   readonly error$ = this.errorSubject.asObservable();
+
+  private scoreChange = new Subject<number>();
+  readonly scoreChange$ = this.scoreChange.asObservable();
 
   private task$ = this.taskInitService.task$.pipe(takeUntil(this.error$));
   private config$ = this.taskInitService.config$.pipe(takeUntil(this.error$));
@@ -78,7 +79,6 @@ export class ItemTaskAnswerService implements OnDestroy {
     private currentAnswerService: CurrentAnswerService,
     private answerTokenService: AnswerTokenService,
     private gradeService: GradeService,
-    private currentContent: CurrentContentService,
   ) {}
 
   ngOnDestroy(): void {
@@ -116,35 +116,12 @@ export class ItemTaskAnswerService implements OnDestroy {
       shareReplay(1),
     );
     combineLatest([ grade$, saveGrade$ ]).subscribe(([ grade ]) => {
-      if (grade.score !== undefined) this.updateCurrentContentScore(grade.score);
+      if (grade.score !== undefined) this.scoreChange.next(grade.score);
     });
     return saveGrade$;
   }
 
   clearAnswer(): Observable<unknown> {
     return this.task$.pipe(take(1), switchMap(task => task.reloadAnswer('')));
-  }
-
-  private updateCurrentContentScore(currentScore: number): void {
-    const content = this.currentContent.current();
-    if (!content) throw new Error('unexpected: no content to update');
-    if (!isItemInfo(content)) throw new Error('unexpected: content is not item info');
-
-    const bestScore = Math.max(content.score?.bestScore ?? 0, currentScore);
-    const isValidated = bestScore === 100;
-    this.currentContent.replace(itemInfo({
-      ...content,
-      details: content.details ? {
-        ...content.details,
-        bestScore,
-        currentScore,
-        validated: isValidated,
-      } : undefined,
-      score: {
-        bestScore,
-        currentScore,
-        isValidated,
-      },
-    }));
   }
 }
