+++ src/app/modules/item/services/item-task.service.ts
@@ -30,6 +30,8 @@ export class ItemTaskService {
   readonly display$ = this.viewsService.display$;
   readonly activeView$ = this.viewsService.activeView$;
 
+  readonly scoreChange$ = this.answerService.scoreChange$;
+
   private navigateToNext$ = this.activityNavTreeService.navigationNeighbors$.pipe(
     map(neighborsState => (neighborsState.isReady ? neighborsState.data?.next?.navigateTo : undefined)),
     shareReplay(1),
